move admin pages to admin directory
admin pages and page parts should be separated from survey pages and page parts, this is a WIP. Renaming pages and components is on the way.
