+++ packages/evolution-frontend/src/components/admin/validations/react-table-config.d.ts
