+++ packages/evolution-legacy/src/components/admin/Validation.js
@@ -12,7 +12,7 @@ import { faFolderOpen } from '@fortawesome/free-solid-svg-icons/faFolderOpen';
 
 import InterviewSummary from './validation/InterviewSummary';
 import { startSetValidateInterview } from 'evolution-frontend/lib/actions/SurveyAdmin';
-import InterviewListComponent from 'evolution-frontend/lib/components/pageParts/validations/InterviewListComponent';
+import InterviewListComponent from 'evolution-frontend/lib/components/admin/validations/InterviewListComponent';
 
 class Validation extends React.Component {
   constructor(props) {
