+++ packages/evolution-legacy/src/components/admin/validation/ValidationOnePageSummary.js
@@ -13,7 +13,7 @@ import appConfig from 'evolution-frontend/lib/config/application.config';
 import config from 'chaire-lib-common/lib/config/shared/project.config';
 import { startUpdateSurveyValidateInterview } from 'evolution-frontend/lib/actions/SurveyAdmin';
 import ValidationCommentForm from './ValidationCommentForm';
-import AdminErrorBoundary from 'evolution-frontend/lib/components/admin/AdminErrorBoundary';
+import AdminErrorBoundary from 'evolution-frontend/lib/components/admin/hoc/AdminErrorBoundary';
 import { generateMapFeatureFromInterview } from 'evolution-frontend/lib/services/admin/odSurveyAdminHelper';
 
 let InterviewStats = appConfig.getCustomInterviewStat();
